+++ client/src/Components/Subscriptions/Member.js
@@ -1,4 +1,4 @@
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import ButtonComp from '../../UI/Button'
 import utils from '../../Utilities/utils';
@@ -7,19 +7,20 @@ import WatchedMoviesComp from '../Movies/WatchedMovies';
 
 const MemberComp = ({ memberData }) => {
 
+  const permissions = useSelector(state => state.usersReducer.connectedUser.permissions)
+
+
   const navigate = useNavigate()
   const dispatch = useDispatch()
 
   const deleteMember = async () => {
-    let status = await utils.deleteItem("http://localhost:5000/api/members", memberData._id)
-
-    if (status.data == "deleted!") {
-
-      dispatch({ type: "DELETE_MEMBER", payload: memberData._id })
-
-      toast("The member was deleted!", { duration: 3000 })
-
-      navigate("/subscriptions/members")
+    if (window.confirm("Are you sure?")) {
+      let status = await utils.deleteItem("http://localhost:5000/api/members", memberData._id)
+      if (status.data == "deleted!") {
+        dispatch({ type: "DELETE_MEMBER", payload: memberData._id })
+        toast("The member was deleted!", { duration: 3000 })
+        navigate("/subscriptions/members")
+      }
     }
   }
 
@@ -34,11 +35,15 @@ const MemberComp = ({ memberData }) => {
       </div>
 
       <div className="actions-box">
-        <ButtonComp type="button" width="20%" height="27px" onClick={() => navigate("/subscriptions/editMember/" + memberData._id)}>Edit</ButtonComp>{" "}
-        <ButtonComp type="button" width="20%" height="27px" onClick={deleteMember} >Delete</ButtonComp>
+        {permissions.find(perm => perm === 'Update Subscriptions') &&
+          <ButtonComp type="button" width="20%" height="27px" onClick={() => navigate("/subscriptions/editMember/" + memberData._id)}>Edit</ButtonComp>
+        }{" "}
+        {permissions.find(perm => perm === 'Delete Subscriptions') &&
+          <ButtonComp type="button" width="20%" height="27px" onClick={deleteMember} >Delete</ButtonComp>
+        }
       </div>
 
-        <WatchedMoviesComp memberID={memberData._id} />
+      <WatchedMoviesComp memberID={memberData._id} />
 
     </div>
   )
